refactor: view 와 domain 패키지 분리
