+++ src/main/java/racingCar/domain/Cars.java
@@ -1,9 +1,10 @@
-package racingCar;
+package racingCar.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+import racingCar.NumberGenerator;
 
 public class Cars {
