+++ src/test/java/racingCar/domain/CarsTest.java
@@ -1,10 +1,11 @@
-package racingCar;
+package racingCar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingCar.NumberGenerator;
 
 @DisplayName("Cars 딘위 테스트")
 class CarsTest {
